requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.46
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.7.2
pydantic[email]==2.8.2
typing-extensions>=4.12.2
cryptography==42.0.5
pycryptodome==3.20.0
psycopg[binary]>=3.0.0